Prevent implicit reading from env vars
Previously, the AWS SDK for Go was configured implicitly by relying on
the environment variables that are present during workflow execution.
The Step did not prepare the input nor configure the SDK in anyway.
This is highly not recommended by Bitrise. Each input that is required
by the Step should be handled and provided explicitly by the user
instead.

This commit declares all the necessary Step input, and uses these input
to configure AWS SDK for Go appropriately.